# secp256k1: Add optimized sqrt field calc.